mv Editor. BoneInfo, BoneSelector, BoneDrawer
