+++ Assets/UniGLTF/Editor/UniGLTF/Gizmo/BoneInfoDrawer.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a5a67f6f96167da4983d24c18f113973
+guid: c6fcbb0dd0018cf4c81922681f91240c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
