+++ Assets/UniGLTF/Editor/UniGLTF/Gizmo/Models/Bone.mat
