+++ Assets/UniGLTF/Editor/UniGLTF/Gizmo/Models/SelectedBone.mat
