--- Assets/UniGLTF/Runtime/UniGLTF/Gizmo/CloverGizmo.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 510b2f063a25b4047a3f275e1c05c410
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
